♻️ 헤더 조건부 렌더링 고차함수로 변환
